refactor : ItemChecker에서 아이템 발송 로직 리팩토링
